Lung X-ray classification/Deployment/lung_mobilenet_hf_space/requirements.txt
fastapi
uvicorn[standard]
tensorflow
numpy
pillow
opencv-python-headless
python-multipart
